Implement a true bias for user priorities as a multiple of default random
Change user priority to a float64 and make sure that it properly biases
the default random priority selection. Using a floating point number
now allows us to set fine grained operation classes (e.g. 1.1x, 1.5x),
as well as set lower operation priority (e.g. 0.5x, 0.1x, 0.01x).

+++ storage/client_replica_test.go
@@ -285,7 +285,7 @@ func TestTxnPutOutOfOrder(t *testing.T) {
 	// priority to trigger the txn restart.
 	manualClock.Increment(100)
 
-	priority := int32(math.MaxInt32)
+	priority := float64(math.MaxInt32)
 	requestHeader := roachpb.Span{
 		Key: roachpb.Key(key),
 	}
